requirements.txt
llama-index
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-llms-anthropic
llama-index-embeddings-voyageai
google-generativeai
openai
anthropic